requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
pandas==2.1.1
numpy==1.24.3
scipy==1.11.3
statsmodels==0.14.0
python-dotenv==1.0.0
openai==2.16.0
